refactored s3/images
