+++ src/services/images/images.service.js
@@ -8,6 +8,7 @@ const BlobService = require('feathers-blob');
 const S3BlobStore = require('s3-blob-store');
 const multer = require('multer');
 const multipartMiddleware = multer();
+const dauria = require('dauria');
 
 var s3 = promise.promisifyAll(new AWS.S3({
   accessKeyId: process.env.S3_ID,
@@ -41,25 +42,47 @@ module.exports = function () {
     // paginate
   };
 
-  // app.use('/images',
-  //   // multer parses the file named 'uri'.
-  //   // Without extra params the data is
-  //   // temporarely kept in memory
-  //   multipartMiddleware.single('uri'),
-  //   // another middleware, this time to
-  //   // transfer the received file to feathers
-  //   function(req,res,next){
-  //     req.feathers.file = req.file;
-  //     next();
-  //   },
-  //   blobService
-  // );
+  app.use('s3/images/new',
+    // multer parses the file named 'uri'.
+    // Without extra params the data is
+    // temporarely kept in memory
+    multipartMiddleware.single('uri'),
+    // another middleware, this time to
+    // transfer the received file to feathers
+    function(req,res,next){
+      req.feathers.file = req.file;
+      next();
+    },
+    blobService
+  );
+
+  app.service('s3/images/new').before({
+    create: [
+      function(hook) {
+        if (!hook.data.uri && hook.params.file){
+          const file = hook.params.file;
+          const uri = dauria.getBase64DataURI(file.buffer, file.mimetype);
+          hook.data = {uri: uri};
+        }
+        const galleryid = hook.data.galleryid ? hook.data.galleryid.toString() : '0';
+        const position = hook.data.position ? hook.data.position.toString() : '0';
+        hook.params.s3 = {
+          ACL: 'public-read',
+          Key: hook.data.name,
+          Metadata: {
+            'galleryid': galleryid,
+            'position': position
+          }
+        };
+      }
+    ]
+  });
 
   // Initialize our service with any options it requires
-  app.use('/images', createService(options));
+  app.use('/s3/images', createService(options));
 
   // Get our initialized service so that we can register hooks and filters
-  const service = app.service('images');
+  const service = app.service('s3/images');
 
   service.hooks(hooks);
